+++ app/assets/javascripts/routers/front/DashboardRouter.js
@@ -212,7 +212,7 @@
       var widgets = this._getDashboardWidgets();
 
       widgets.forEach(function (widget, index) {
-        if (widget && widget.type === 'map') {
+        if (widget && widget.type === 'map' && widget.visible) {
           this['widget' + index] = new App.View.MapWidgetView({
             el: document.querySelector('.js-widget-' + (index + 1)),
             data: dataset,
@@ -226,13 +226,16 @@
             center: [widget.lat || 0, widget.lng || 0],
             zoom: widget.zoom
           });
-        } else if (widget && widget.type === 'chart') {
+        } else if (widget && widget.type === 'chart' && widget.visible) {
           this['widget' + index] = new App.View.ChartWidgetView({
             el: document.querySelector('.js-widget-' + (index + 1)),
             data: dataset,
             chart: widget.chart,
             columnX: widget.x,
-            columnY: widget.y
+            columnY: widget.y,
+            xLabel: widget.xLabel,
+            yLabel: widget.yLabel,
+            displayMode: 'dashboard'
           });
         }
       }, this);
@@ -282,15 +285,18 @@
         f: state.config.filters.map(function (filter) { return [filter.name, filter.value]; }),
         w: this.state.config.widgets.map(function (chart) {
           if (!chart.type) return null;
-          var res = { t: chart.type };
+          var res = {
+            t: chart.type,
+            v: chart.visible
+          };
 
           if (chart.type === 'map') {
             res.la = chart.lat;
             res.ln = chart.lng;
             res.z = chart.zoom;
           } else {
-            res.c = chart.chart;
             res.x = chart.x;
+            res.c = chart.chart;
             if (chart.y) res.y = chart.y;
           }
 
@@ -335,10 +341,11 @@
           widgets: compressedState.w && compressedState.w.map(function (widget) {
             if (widget.t === 'map') {
               return {
-                type: 'map',
+                type: 'chart',
                 lat: widget.la,
                 lng: widget.ln,
-                zoom: widget.z
+                zoom: widget.z,
+                visible: widget.v
               };
             }
 
@@ -346,7 +353,8 @@
               type: 'chart',
               chart: widget.c,
               x: widget.x,
-              y: widget.y || null
+              y: widget.y || null,
+              visible: widget.v || true
             };
           })
         }
@@ -430,7 +438,7 @@
       var dataset = this._getDataset({ unfiltered: true });
 
       // We check we have all the widgets and the dataset isn't empty
-      if (!state.config || !state.config.widgets || state.config.widgets.length < this.options.widgetsCount || !dataset.length) {
+      if (!state.config || !state.config.widgets || !dataset.length) {
         return false;
       }
 
@@ -448,7 +456,7 @@
 
     /**
      * Restore the state of the dashboard
-     * NOTE: must be called after _renderWidgets
+     * NOTE: must be called before _renderWidgets
      * @param {object} state
      * @returns {boolean} restored - true if the state could be restored
      */
@@ -478,17 +486,19 @@
       // We restore the widgets
       for (var i = 0, j = this.options.widgetsCount; i < j; i++) {
         var widget = state.config.widgets[i];
-        if (widget.type === 'map') {
-          this['widget' + i].options = Object.assign({}, this['widget' + i].options, {
-            center: [widget.lat, widget.lng],
-            zoom: widget.zoom
-          });
-        } else {
-          this['widget' + i].options = Object.assign({}, this['widget' + i].options, {
-            chart: widget.chart,
-            columnX: widget.x,
-            columnY: widget.y
-          });
+        if (this['widget' + i]) {
+          if (widget && widget.type === 'map') {
+            this['widget' + i].options = Object.assign({}, this['widget' + i].options, {
+              center: [widget.lat, widget.lng],
+              zoom: widget.zoom
+            });
+          } else if (widget){
+            this['widget' + i].options = Object.assign({}, this['widget' + i].options, {
+              chart: widget.chart,
+              columnX: widget.x,
+              columnY: widget.y
+            });
+          }
         }
       }
 
@@ -503,7 +513,7 @@
      */
     _renderWidgets: function () {
       for (var i = 0, j = this.options.widgetsCount; i < j; i++) {
-        if(this['widget' + i]) this['widget' + i].render();
+        if (this['widget' + i]) this['widget' + i].render();
       }
     },