+++ app/assets/javascripts/views/shared/chartWidgetView.js
@@ -26,7 +26,9 @@
       // Inner width of the chart, used internally
       _width: null,
       // Inner height of the chart, used internally
-      _height: null
+      _height: null,
+      // Flag that determines whether the widget is visible or not
+      visible: true
     },
 
     initialize: function (settings) {
@@ -72,6 +74,9 @@
       this.options.chart = arguments[0][0];
       this.options.columnX = arguments[0][1];
       this.options.columnY = arguments[0].length > 2 ? arguments[0][2] : null;
+      this.options.xLabel = null;
+      this.options.yLabel = null;
+      this._renderCustomAxisLabelInput();
       this._renderChart();
     },
 
@@ -85,6 +90,8 @@
         data: JSON.stringify([]),
         xColumn: JSON.stringify(''),
         yColumn: JSON.stringify(''),
+        xLabel: JSON.stringify(''),
+        yLabel: JSON.stringify(''),
         width: JSON.stringify(''),
         height: JSON.stringify('')
       }));
@@ -154,12 +161,28 @@
         this.options.columnX = columns.x;
         this.options.columnY = columns.y;
       }
+
+      const columnX = JSON.stringify(this.options.columnX);
+      const columnY = JSON.stringify(this.options.columnY);
+      const labelX = this.options.xLabel;
+      const labelY = this.options.yLabel;
+      var xLabel = columnX;
+      var yLabel = columnY;
+
+      if (labelX !== undefined && labelX !== null && labelX !== ''){
+        xLabel = JSON.stringify(labelX);}
+
+      if (labelY !== undefined && labelY !== null && labelY !== ''){
+        yLabel = JSON.stringify(labelY);}
+
       return this._getChartTemplate()({
         data: JSON.stringify(this.options.data),
-        xColumn: JSON.stringify(this.options.columnX),
-        yColumn: JSON.stringify(this.options.columnY),
+        xColumn: columnX,
+        yColumn: columnY,
         width: chartDimensions.width,
-        height: chartDimensions.height
+        height: chartDimensions.height,
+        xLabel: xLabel,
+        yLabel: yLabel
       });
     },
 
@@ -181,7 +204,7 @@
           }
           this.chart = chart({
             el: this.chartContainer,
-            // By using the SVG renderer, we give the client the posibility to
+            // By using the SVG renderer, we give the client the possibility to
             // translate the text contained in the charts
             renderer: 'svg'
           }).update();
@@ -192,13 +215,16 @@
 
       // We save the state of the widget each time we render as it can be the
       // consequence of a change in the configuration
-      // NOTE: We need to make sure in case the view hasn't been instanciated with
+      // NOTE: We need to make sure in case the view hasn't been instantiated with
       // a chart configuration that it's set by the toolbox before triggering
       this.trigger('state:change', {
         type: 'chart',
         chart: this.options.chart,
         x: this.options.columnX,
-        y: this.options.columnY
+        y: this.options.columnY,
+        xLabel: this.options.xLabel,
+        yLabel: this.options.yLabel,
+        visible: this.options.visible
       });
     },
 
@@ -259,7 +285,66 @@
       // We append the button to the DOM
       this.el.querySelector('.js-switch-button').appendChild(button);
     },
+    /**
+     * Render the Toggle Visibility button and attach an event listener to it
+     */
+    _renderToggleVisibilityButton: function () {
+      // We create the button
+      var button = document.createElement('button');
+      button.type = 'button';
+      button.classList.add('c-button', 'toggle-visibility-button');
+      if (!this.options.visible) button.classList.add('-slashed');
+      button.textContent = 'Hide';
+      // We attach the listener
+      button.addEventListener('click', function() {
+        button.classList.toggle('-slashed');
+        this._toggleVisibility();
+      }.bind(this));
 
+      // We append the button to the DOM
+      this.el.querySelector('.js-toggle-visibility-button').appendChild(button);
+    },
+    /**
+     * Trigger widget visibility change
+     */
+    _toggleVisibility: function () {
+      this.options.visible = !this.options.visible;
+      this.trigger('state:change', { visible: this.options.visible });
+    },
+    /**
+     * Render the Custom Axis Label Inputs and attach an event listener to it
+     */
+    _renderCustomAxisLabelInput: function () {
+      var inputContainer = this.el.querySelector('#custom-axis-input-container');
+      if (inputContainer.children.length) {
+        inputContainer.innerHTML = '';
+      }
+
+      const axis = ['X','Y'];
+
+      axis.forEach(function(axis){
+        // We create the input
+        var input = document.createElement('input');
+        input.setAttribute('placeholder', 'Custom ' + axis + ' label');
+        input.setAttribute('name', 'custom-' + axis);
+        const axisLabel = (axis === 'X') ? this.options.xLabel : this.options.yLabel;
+        if (axisLabel !== undefined) { input.value = axisLabel };
+
+        // We attach the listeners
+        input.addEventListener('change', function(){
+          this._changeAxisLabel(axis , input.value);
+        }.bind(this));
+
+        // We append the inputs to the DOM
+        inputContainer.appendChild(input);
+
+      }.bind(this))
+    },
+    _changeAxisLabel: function(axis, value){
+      if(axis === 'X') this.options.xLabel = value;
+      else this.options.yLabel = value;
+      this._renderChart();
+    },
     /**
      * Remove the changes the component implied to the container and all of
      * its children
@@ -273,9 +358,16 @@
 
       // We don't render the chart selector and the switch button if the dataset is empty
       if (this.options.data.length) {
-        if (this.options.enableChartSelector) this._renderChartSelector();
+        if (this.options.enableChartSelector){
+          this._renderChartSelector();
+          // TODO: use displayMode to enable the render of rest of the buttons and inputs
+          if(this.options.displayMode !== 'dashboard'){
+            this._renderCustomAxisLabelInput();
+          }
+        }
         if (this.options.switchCallback && typeof this.options.switchCallback === 'function') {
           this._renderSwitchButton();
+          this._renderToggleVisibilityButton();
         }
       }