+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+### 4 April 2017
+- Add custom labels support to widgets
 - Added wysiwyg css cheatsheet
 - Changes wysiwyg formatting styles
 - Added debugging to custom router to identify potential issue when rebuilding
-
+- Terms of Service page update
+- Adds visibility toggle to dashboard widgets
 
 ### 29 Mar 2017
 - Add validation of template selection to site creation